requirements.txt
scikit-learn
matplotlib
tk
io
pickle5
Pillow
numpy